// case_study_angular/CaseStudy/src/app/caseStudy/model/Facility.ts
export interface Facility {
  facilityId?: number;
  facilityName?: string;
  facilityArea?: number;
  rentCost?: number;
  maxPeople?: number;
  standardRoom?: string;
  descriptionOtherConvenience?: string;
  poolArea?: number;
  numberOfFloors?: number;
  facilityFree?: string;
  rentType?: string;
  facilityType?: string;
  facilityImage?: string;

}